backend/requirements.txt
blinker==1.8.2
boto3==1.35.14
botocore==1.35.14
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jose==1.0.0
jwt==1.3.1
MarkupSafe==2.1.5
packaging==24.1
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
SQLAlchemy==2.0.34
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
